requirements.txt
flask
requests
selenium
pandas
python-binance
python-dotenv
